+++ src/main/java/com/codetaylor/mc/pyrotech/modules/hunting/block/fluid/BlockFluidTannin.java
@@ -1,0 +1,15 @@
+package com.codetaylor.mc.pyrotech.modules.hunting.block.fluid;
+
+import net.minecraft.block.material.Material;
+import net.minecraftforge.fluids.BlockFluidClassic;
+import net.minecraftforge.fluids.Fluid;
+
+public class BlockFluidTannin
+    extends BlockFluidClassic {
+
+  public BlockFluidTannin(Fluid fluid) {
+
+    super(fluid, Material.WATER);
+  }
+
+}